+++ catalog/app/containers/Admin/Buckets/Buckets.tsx
@@ -21,6 +21,7 @@ import type FormSpec from 'utils/FormSpec'
 import * as GQL from 'utils/GraphQL'
 import MetaTitle from 'utils/MetaTitle'
 import * as NamedRoutes from 'utils/NamedRoutes'
+import StyledLink from 'utils/StyledLink'
 import StyledTooltip from 'utils/StyledTooltip'
 import assertNever from 'utils/assertNever'
 import parseSearch from 'utils/parseSearch'
@@ -766,6 +767,12 @@ interface PrimaryCardProps {
 function PrimaryCard({ bucket, className, disabled, onSubmit }: PrimaryCardProps) {
   const initialValues = bucketToPrimaryValues(bucket)
   const ref = React.useRef<HTMLElement>(null)
+  const { urls } = NamedRoutes.use()
+  const configPath = '.quilt/catalog/config.yaml'
+  const configHref = urls.bucketFile(bucket.name, configPath, {
+    edit: true,
+  })
+  const configUrl = `s3://${bucket.name}/${configPath}`
   return (
     <RF.Form<PrimaryFormValues> onSubmit={onSubmit} initialValues={initialValues}>
       {({ handleSubmit, form, submitFailed }) => (
@@ -780,7 +787,15 @@ function PrimaryCard({ bucket, className, disabled, onSubmit }: PrimaryCardProps
             <PrimaryForm bucket={bucket} />
           </form>
           <StickyActions parentRef={ref}>
-            <CardActions<PrimaryFormValues> disabled={disabled} form={form} />
+            <CardActions<PrimaryFormValues>
+              action={
+                <StyledLink to={configHref}>
+                  Modify bucket appearance via {configUrl}
+                </StyledLink>
+              }
+              disabled={disabled}
+              form={form}
+            />
           </StickyActions>
         </Card>
       )}